+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 * [MIC-6835] Add New Card Screen Redisign - Present Add Card Only Integration
 * [MIC-6836] Add New Card Screen Redisign - Sample add shortcut button for presenting Add Card
 * [MIC-7538] Add New Tinkoff Bins
+* [MIC-7545] Adapt CardList for Payment Card Selection Flow
 
 ### Fixed